+++ engine/Display/EasterEgg.cpp
@@ -10,8 +10,8 @@
 #endif
 
 void EasterEgg::CunningBlit
-	(uint16* from, 
-	 uint16* to, 
+	(uint32* from, 
+	 uint32* to, 
 	 int width, 
 	 int height, 
 	 int fromPitch,
